requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.10.3
backoff==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.4
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.9.24
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.6
cloud-init==24.1.3
colorama==0.4.6
colorzero==2.0
command-not-found==0.3
configobj==5.0.8
contourpy==1.2.0
cryptography==42.0.7
cupshelpers==1.0
cvzone==1.6.1
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
distro==1.8.0
distro-info==1.5+ubuntu0.23.10.1
dnspython==2.6.1
dominate==2.8.0
duplicity==1.2.2
et-xmlfile==1.1.0
exotel==0.1.5
fasteners==0.17.3
filelock==3.13.3
firebase-admin==6.2.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-crontab==0.1.2
Flask-Datepicker==0.14
Flask-Ext==0.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.4
flatbuffers==24.3.25
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2024.3.1
future==0.18.2
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gpiozero==2.0.1
greenlet==3.0.0
grpcio==1.59.0
grpcio-status==1.59.0
gyp==0.1
h11==0.14.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jax==0.4.25
jaxlib==0.4.25
jeepney==0.8.0
Jinja2==2.11.2
JSON-log-formatter==0.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
jws==0.1.3
keyring==24.2.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libevdev==0.5
lockfile==0.12.2
louis==3.26.0
lxml==5.2.2
Mako==1.2.4.dev0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
ndg-httpsclient==0.5.1
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
onfleet==0.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
paramiko==2.12.0
pexpect==4.8.0
Pillow==10.0.0
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycairo==1.24.0
pycparser==2.22
pycryptodome==3.19.0
pycups==2.0.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
pymongo==4.6.3
PyMsgBox==1.0.9
PyMuPDF==1.23.20
PyMuPDFb==1.23.9
PyNaCl==1.5.0
pyonfleet==1.4.0
pyOpenSSL==24.1.0
pyparsing==3.1.0
PyPDF2==3.0.1
pyperclip==1.8.2
Pyrebase==3.0.27
Pyrebase4==4.7.1
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-docx==1.1.2
python-dotenv==1.0.0
python-jwt==2.0.1
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.2.0
pytz==2023.3
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
ratelimit==2.2.1
razorpay==1.4.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.3.1
rsa==4.9
scipy==1.12.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
selenium==4.15.2
shortuuid==1.0.11
six==1.16.0
slack==0.0.2
slackclient==2.9.4
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==1.4.0
sympy==1.12
systemd-python==235
terminator==2.1.3
torch==2.2.2
trio==0.23.1
trio-websocket==0.11.1
triton==2.2.0
typing_extensions==4.12.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.17
usb-creator==0.3.16
visitor==0.1.3
wadllib==1.3.6
webdriver-manager==4.0.1
Werkzeug==2.0.3
WooCommerce==3.0.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
xml-to-dict==0.1.6
xmltodict==0.13.0
yarl==1.9.2
zipp==1.0.0
